// src/app/modules/users/models/persona.model.ts
export interface PersonaModel {
    id: number;
    primerNombre: string;
    primerApellido: string;
    segundoNombre: string;
    segundoApellido: string;
    fechaNacimiento: Date;
    identificacion: number;
    direccion: string;
    telefono: number;
    idGenero: number;
    uid: number;
    idPlaneta: number;
    edad: number;
    peso: number;
    contador: number;
}